# Documentazione/OpenAPI.yaml
openapi: "3.0.2"
info:
  title: "Aulery"
  description: "Specifica progetto Aulery"
  version: "1.0"
  termsOfService: "https://aulery.it/terms"
  contact:
    name: "Aulery API by Ansima"
    url: "https://aulery.it/info"
    email: "contact@aulery.it"
  license:
    name: "CC Aulery1.0(CC AU 1.0)"
    url: "https://aulery.it"
...
servers:
  - url: "localhost/api/"
    description: Server LocalHost
components:
  schemas:
...
    Area:
      type: object
      properties:
        nome:
          type: string
        descrizione:
          type: string
    Corso:
      type: object
      properties:
        nome:
          type: string
        area:
          $ref: '#/components/schemas/Area'
    Evento:
      type: object
      properties:
        nome:
          type: string
        descrizione:
          type: string
        data:
          type: string
          format: date
        oraInizio:
          type: string
          format: time
        oraFine:
          type: string
          format: time
        tipologiaEvento:
          $ref: '#/components/schemas/TipologiaEvento'
        persona:
          $ref: '#/components/schemas/Persona'
        aula:
          $ref: '#/components/schemas/Aula'
        corso:
          $ref: '#/components/schemas/Corso'
        eventoRicorrente:
          $ref: '#/components/schemas/EventoRicorrente'
    EventoRicorrente:
      type: object
      properties:
        dataFine:
          type: string
          format: date
        ricorrenza:
          $ref: '#/components/schemas/Ricorrenza'
    Aula:
      type: object
      properties:
        nome:
          type: string
        note:
          type: string
        edificio:
          $ref: '#/components/schemas/Edificio'
        piano:
          type: integer
        zona:
          type: string
        capienza:
          type: integer
        preseElettriche:
          type: integer
        preseDiRete:
          type: integer
        attrezzature:
          type: array
          items:
            $ref: '#/components/schemas/Attrezzatura'
        area:
          $ref: '#/components/schemas/Area'
    Persona:
      type: object
      properties:
        nome:
          type: string
        cognome:
          type: string
        email:
          type: string  
    Attrezzatura:
      type: string
      enum:
        - PROIETTORE
        - SCHERMOMOTORIZZATO
        - SCHERMOMANUALE
        - IMPIANTOAUDIO
        - PCFISSO
        - MICROFONOAFILO
        - MICROFONOSENZAFILO
        - LAVAGNALUMINOSA
        - WIFI
    Edificio:
      type: string
      enum:
        - COPPITO0
        - COPPITO1
        - COPPITO2
        - ALANTURING
        - BLOCCOA
        - BLOCCOB
        - BLOCCOC
    Ricorrenza:
      type: string
      enum:
        - GIORNALIERA
        - SETTIMANALE
        - MENSILE
        - ANNUALE  
    TipologiaEvento:
      type: string
      enum:
        - LEZIONE
        - ESERCITAZIONE
        - ESAME
        - PARZIALE
        - LAUREE
        - SEMINARIO
        - CONFERENZA
        - RIUNIONE
        - CDC
        - CAD
        - CERIMONIA
        - ELEZIONI
        - DIPARTIMENTO
        - ATENEO
        - ESTERNO
        - MANUTENZIONE
        - STUDIOINDIVIDUALE
        - ALTRO
  parameters:
...
    areaID:
      name: areaID
      in: path
      description: Id dell'area
      required: true
      schema:
        type: integer
    eventoID:
      name: eventoID
      in: path
      description: Id dell'evento
      required: true
      schema:
        type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: internalToken
  responses:
    GeneralError:
      description: Errore 500, Errore interno
      content:
        text/plain:
          schema:
            type: string
            example: "Errore 500, Errore interno"
...
    NotFoundError:
      description: Errore 404, Elemento non trovato
      content:
        text/plain:
          schema:
            type: string
            example: "Errore 404, Elemento non trovato"   
paths:
...
  /rest/eventi/crea:
    post:
      summary: Crea evento
      description: Operazione che permette di aggiungere un evento al sistema.
      operationId: creaEvento
      tags: 
        - Eventi
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evento'
      responses:
        201:
          description: Evento creato con successo
          headers:
            location:
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                type: string
                format: uri
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/GeneralError'
  /rest/eventi/{eventoID}/modifica:
    put:
      summary: Modfica evento
      description: Operazione che permette di modificare una evento già presente nel sistema.
      operationId: updateEvento
      tags: 
        - Eventi
      security:
        - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/eventoID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evento'
      responses:
        204:
          description: no content
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/GeneralError'
  /rest/eventi/{eventoID}/info:
    get:
      summary: Info evento
      description: Operazione che permette di ottenere le informazioni riguardo un evento già presente nel sistema.
      operationId: getEvento
      tags:
        - Eventi
      parameters:
      - $ref: '#/components/parameters/eventoID'
      responses:
        '200':
          description: Informazioni ottenute con successo 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evento'
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/GeneralError' 
...
  /rest/collezioni/{areaID}/prossime3ore:
    get:
      summary: Lista eventi delle prossime tre ore
      description: Operazione che permette di ottenere una lista di eventi attuali e delle prossime tre ore di una intera area.
      operationId: getEventiTreOre
      tags: 
        - Liste
      parameters:
      - $ref: '#/components/parameters/areaID'
      responses:
        200:
          description: Lista ottenuta con successo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evento'
        404:
          $ref: '#/components/responses/NotFoundError'  
        500:
          $ref: '#/components/responses/GeneralError'